# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "wkwkwkwkwk description",
        "title": "CRUD with Gin Golang",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "saya",
            "url": "http://github.com/haynafi"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/",
    "paths": {
        "/product/{id}": {
            "get": {
                "description": "get product by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Show"
                ],
                "summary": "Show an product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "description": "insert product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Insert"
                ],
                "summary": "Insert product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nama Produk",
                        "name": "nama_product",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "deskripsi",
                        "name": "deskripsi",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/products": {
            "get": {
                "description": "get all product",
                "tags": [
                    "Get"
                ],
                "summary": "Show all product",
                "responses": {}
            }
        }
    }
}